Squashed the dev/scratch commits into an initial release commit. More details listed below:
TriangleMesh.h contains code for generating Triangle Meshes.
ClosestPointQuery.h implements an interface for finding the closest point on a mesh to an external query point.
TestSuite.h provides two basic tests: One for testing if our output matches that of CGAL; and another to test if multiple point queries on the same mesh are efficient.
Common.h contains includes headers and typedefs required for our classes.
There is only one compilation unit.

NOTES will have to be updated.
.gitignore might need to have some files added.
The program behaviour is untested or undefined in some cases, such as if the query point is on the mesh.